BARK-57: Move opGetVetClinics functionality into new VetAccountService class
