+++ src/lib/admin/rpc/commands/list-vet-clinics.ts
@@ -1,5 +1,5 @@
+import APP from "@/lib/app";
 import { AdminCommand, AdminCommandArgs } from "../admin-command";
-import { opGetVetClinics } from "@/lib/bark/operations/op-get-vet-clinics";
 
 export class ListVetClinics implements AdminCommand {
   getExampleRequest(): string {
@@ -7,8 +7,9 @@ export class ListVetClinics implements AdminCommand {
   }
 
   async run(args: AdminCommandArgs): Promise<string> {
-    const { context } = args;
-    const { result, error } = await opGetVetClinics(context);
+    // WIP: Should call actor.getVetClinics()
+    const service = await APP.getVetAccountService();
+    const { result, error } = await service.getVetClinics();
     if (error) {
       throw new Error(error);
     }
